requirements.txt
annotated-types==0.6.0
anyio==4.3.0
certifi==2024.2.2
distro==1.9.0
exceptiongroup==1.2.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
numpy==1.24.4
openai==0.27.10
pillow==10.3.0
pydantic==2.7.1
pydantic-core==2.18.2
PyMuPDF==1.24.4
PyMuPDFb==1.24.3
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
sniffio==1.3.1
tqdm==4.66.4
typing-extensions==4.11.0
tzdata==2024.1
